refactor: event/views.py を機能別モジュールに分割
1737行の views.py を10個のモジュールに分割し、可読性と保守性を向上。
__init__.py で全ビューを re-export し、既存の urls.py や他モジュールからの
インポートパスを維持する。テストの mock パッチ先も新しいモジュールパスに更新。

分割構成:
- helpers.py: 権限判定・YouTube URL解析ヘルパー
- list.py: EventListView, EventDetailPastList, EventLogListView
- my_list.py: EventMyList (管理画面)
- detail.py: EventDetailView
- crud.py: EventDeleteView, EventDetailCreateView/UpdateView/DeleteView
- calendar_create.py: GoogleCalendarEventCreateView
- blog.py: GenerateBlogView, BigQuery関連
- sync.py: Googleカレンダー同期関数
- lt_application.py: LT申請関連ビュー4種

Closes #198
