+++ app/event/tests/test_bigquery_lazy_init.py
@@ -17,7 +17,7 @@ class TestBigQueryLazyInit(TestCase):
         """
         # views.pyがインポート済みの場合でも、グローバル変数が
         # Noneのままであることを確認
-        from event.views import _bigquery_client, _bigquery_project
+        from event.views.blog import _bigquery_client, _bigquery_project
 
         # _get_bigquery_clientが呼ばれるまでNoneのまま
         self.assertIsNone(_bigquery_client)
@@ -27,12 +27,12 @@ class TestBigQueryLazyInit(TestCase):
     def test_get_bigquery_client_returns_mock_in_testing(self):
         """TESTING環境変数が設定されている場合はMockを返すこと。"""
         # キャッシュをリセットするためにグローバル変数をリセット
-        import event.views as views_module
-        views_module._bigquery_client = None
-        views_module._bigquery_project = None
+        import event.views.blog as blog_module
+        blog_module._bigquery_client = None
+        blog_module._bigquery_project = None
 
         try:
-            from event.views import _get_bigquery_client
+            from event.views.blog import _get_bigquery_client
 
             client, project = _get_bigquery_client()
 
@@ -41,18 +41,18 @@ class TestBigQueryLazyInit(TestCase):
             self.assertIsNotNone(client)
         finally:
             # テスト後にクリーンアップ
-            views_module._bigquery_client = None
-            views_module._bigquery_project = None
+            blog_module._bigquery_client = None
+            blog_module._bigquery_project = None
 
     @patch.dict(os.environ, {'TESTING': '1'})
     def test_get_bigquery_client_is_cached(self):
         """_get_bigquery_clientが結果をキャッシュすること。"""
-        import event.views as views_module
-        views_module._bigquery_client = None
-        views_module._bigquery_project = None
+        import event.views.blog as blog_module
+        blog_module._bigquery_client = None
+        blog_module._bigquery_project = None
 
         try:
-            from event.views import _get_bigquery_client
+            from event.views.blog import _get_bigquery_client
 
             # 1回目の呼び出し
             client1, project1 = _get_bigquery_client()
@@ -64,5 +64,5 @@ class TestBigQueryLazyInit(TestCase):
             self.assertEqual(project1, project2)
         finally:
             # テスト後にクリーンアップ
-            views_module._bigquery_client = None
-            views_module._bigquery_project = None
+            blog_module._bigquery_client = None
+            blog_module._bigquery_project = None
