+++ app/event/tests/test_event_detail_permissions.py
@@ -205,7 +205,7 @@ class EventDetailPermissionTests(TestCase):
         self.assertEqual(self.applicant_detail.theme, "Updated Theme")
         self.assertTrue(bool(self.applicant_detail.slide_file))
 
-    @patch("event.views.generate_blog")
+    @patch("event.views.blog.generate_blog")
     def test_applicant_can_generate_blog_for_approved_event_detail(self, mock_generate_blog):
         """発表者本人は自分の承認済みLTで記事生成できる."""
         self.client.login(username="applicant_user", password="testpass123")
@@ -230,7 +230,7 @@ class EventDetailPermissionTests(TestCase):
         self.assertEqual(self.applicant_detail.meta_description, "生成ディスクリプション")
         self.assertEqual(self.applicant_detail.contents, "生成本文")
 
-    @patch("event.views.generate_blog")
+    @patch("event.views.blog.generate_blog")
     def test_applicant_cannot_generate_blog_for_pending_event_detail(self, mock_generate_blog):
         """発表者本人でも承認待ちLTでは記事生成できない."""
         self.client.login(username="applicant_user", password="testpass123")
